requirements.txt
gcodeparser
numpy
plotly
pyyaml
scipy
trimesh
